pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "srm-svm-steganalysis"
version = "0.1.0"
description = "Sparse Representation Model with SVM for Image Steganalysis"
authors = [{name = "Author", email = "author@example.com"}]
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "scikit-learn>=1.0.0",
    "matplotlib>=3.5.0",
    "Pillow>=8.3.0",
    "joblib>=1.1.0",
    "tqdm>=4.62.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.2.0",
    "mypy>=0.910",
    "flake8>=4.0.0",
    "black>=22.0.0",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", "build", "dist"]

[tool.black]
line-length = 88
target-version = ["py310"]